requirements.txt.txt
transformers==4.0.0
python-telegram-bot==13.0